Fixed data types issues (loss due to data types conversion), initializations and typos

+++ libdash/libdash/include/IEvent.h
@@ -41,7 +41,7 @@ namespace dash
 				 *
                  *  @return     an unsigned integer
                  */
-                virtual uint32_t             GetPresentationTime    ()  const = 0;
+                virtual uint64_t             GetPresentationTime    ()  const = 0;
 
                 /**
                  *  Returns a reference to a string that specifies the presentation duration of the Event. \n
